long

+++ segmenter_module.py
@@ -33,7 +33,7 @@ class SegmenterModule(pl.LightningModule):
         image, depth, label = batch['image'], batch['depth'], batch['label']
 
         y_hat = self(image)
-        loss = F.cross_entropy(y_hat, label)
+        loss = F.cross_entropy(y_hat, label.long())
         acc = torchmetrics.functional.iou(
             y_hat.clone().detach(), label, num_classes=self.num_classes)
 
@@ -45,7 +45,7 @@ class SegmenterModule(pl.LightningModule):
         image, depth, label = batch['image'], batch['depth'], batch['label']
 
         y_hat = self(image)
-        loss = F.cross_entropy(y_hat, label)
+        loss = F.cross_entropy(y_hat, label.long())
         acc = torchmetrics.functional.iou(
             y_hat.clone().detach(), label, num_classes=self.num_classes)
 
@@ -57,7 +57,7 @@ class SegmenterModule(pl.LightningModule):
         image, depth, label = batch['image'], batch['depth'], batch['label']
 
         y_hat = self(image)
-        loss = F.cross_entropy(y_hat, label)
+        loss = F.cross_entropy(y_hat, label.long())
         acc = torchmetrics.functional.iou(
             y_hat.clone().detach(), label, num_classes=self.num_classes)
